+++ src/middleware/auth/key-auth.ts
