+++ src/middleware/authentication.ts
@@ -8,7 +8,7 @@ import { configLogToExpress } from '../types/constants.js'
 import { AccountAuthHandler } from './auth/account-auth.js'
 import { CredentialAuthHandler } from './auth/credential-auth.js'
 import { DidAuthHandler } from './auth/did-auth.js'
-import { KeyAuthHandler } from './auth/key_auth.js'
+import { KeyAuthHandler } from './auth/key-auth.js'
 import { CredentialStatusAuthHandler } from './auth/credential-status-auth.js'
 import { AbstractAuthHandler } from './auth/base-auth.js'
 import { LogToHelper } from './auth/logto.js'
